feat(per): Display flags from Person Escort Record on move detail
This introduces the ability to display a tag list from the Person
Escort Record framework if this feature is enabled.

It will replace the existing method of displaying the tags from the
assessment answers, but for now both need to live side by side until
the feature is completely released.