feat(per): Display card tags from Person Escort Record
This updates the profile to card component to display tags based
on the Person Escort Record instead of the previous assessment answers.

Because the feature is being developed in the background to avoid any
breaking changes this change needs to support the existing method unless
a feature flag is enabled.